Fixing the unit tests
